pyproject.toml
[project]
name = "poppulo-rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.38.23",
    "pip>=25.1.1",
    "python-dotenv>=1.1.0",
    "spacy>=3.8.7",
    "pinecone>=7.0.1"
]
